Clear the -Wfloat-conversion warnings; pass third-party includes as -isystem
The flag added in 0.9.1 reported 60 lines on a debug build: 48 from STRUMPACK
headers under $SCLS/include, 12 from BELFEM sources.

Build: finalize_compiler.cmake now passes every BELFEM_INCLUDES directory to
the C++ compiler as -isystem (normalized, deduplicated), so a warning raised
inside a vendor header is suppressed and cannot trip -Wall -Werror. A directory
already in CMAKE_CXX_IMPLICIT_INCLUDE_DIRECTORIES keeps -I, so /usr/include is
never reordered ahead of the standard library. The per-test BEFORE include of
$SCLS/include in Add_Test.cmake was redundant in both $SCLS branches and is
removed; tests now search BELFEM's headers before SCLS, as the library did.
The Fortran loop keeps -I. -fallow-argument-mismatch is documented as
deliberate (PARDISO's F77 scalar-dummy idiom is an error on gfortran >= 10).

Sources: the 12 reported sites, plus two of the same class in cl_Genome.hpp
that only the test build compiles, are explicit casts of the same expressions;
<cmath> is included where ceil/round had been unqualified.
spline::create_helpmatrix takes its point count as const index_t and its
spacing as const real, both by value (was const real & for both); the four
non-index_t callers cast explicitly. Public API change; consumers recompile.

Input validation (Christian's ruling, setup-tier): Section::get_int rounds and
then BELFEM_ERRORs on a non-finite value or one outside int; a unit exponent
is parsed with to_real() and must be an integer with |e| <= 32, so m^1.5 and
m^abc stop with an error instead of truncating or escaping as an exception.
New test StringTools.UnitToSiExponentGuard.

Blind plan and code juries (Codex gpt-5.6-terra, Grok 4.6) refuted three plan
rows as not behavior-preserving (stoi vs stod, cast moved across +1, INT_MIN
bound) and a false 0.9.0 attribution in the first comment; all corrected.
Nine library TUs and the test TU compile warning-free under their module
flags with -fsyntax-only. Not built, not run: the reconfigure, make and
make check are the gate. No peer work rides along; CHANGELOG.md and
devlog/README.md are staged hunk-only, their concurrent entries stay in the
working tree.

+++ src/containers/cl_Genome.hpp
@@ -12,6 +12,7 @@
 #ifndef BELFEM_CL_DNA_HPP
 #define BELFEM_CL_DNA_HPP
 
+#include <cmath>
 #include <algorithm>
 
 #include "typedefs.hpp"
@@ -75,7 +76,7 @@ namespace belfem
                 tValue = ( tValue - tMin ) / ( tMax - tMin );
 
                 // scale to [0, 2^B - 1] and encode
-                encode( i, std::round( tValue * ( ( 1u << B ) - 1 ) ) );
+                encode( i, static_cast< uint >( std::round( tValue * ( ( 1u << B ) - 1 ) ) ) );
             }
         }
 
@@ -169,7 +170,7 @@ namespace belfem
                 tValue = ( tValue - tMin ) / ( tMax - tMin );
 
                 // encode
-                encode( i, std::round( tValue * ( ( 1u << B ) - 1 ) ) );
+                encode( i, static_cast< uint >( std::round( tValue * ( ( 1u << B ) - 1 ) ) ) );
             }
         }
